Trait pliantdb_server::Backend

pub trait Backend: Debug + Send + Sync + Sized + 'static {
    type CustomApi: CustomApi;
    type CustomApiDispatcher: Dispatcher<<Self::CustomApi as CustomApi>::Request, Result = Result<<Self::CustomApi as CustomApi>::Response>> + Debug;
#[must_use]    fn client_connected<'async_trait>(
        client: ConnectedClient<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
    { ... }

#[must_use]    fn client_disconnected<'async_trait>(
        client: ConnectedClient<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
    { ... }
}

Tailors the behavior of a server to your needs.

Associated Types

type CustomApi: CustomApi

The custom API definition. If you do not wish to have an API, () may be provided.

type CustomApiDispatcher: Dispatcher<<Self::CustomApi as CustomApi>::Request, Result = Result<<Self::CustomApi as CustomApi>::Response>> + Debug

The type that implements the Dispatcher trait.

Loading content...
Provided methods

#[must_use]fn client_connected<'async_trait>(
client: ConnectedClient<Self>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
Self: 'async_trait,

A client disconnected from the server. This is invoked before authentication has been performed.

#[must_use]fn client_disconnected<'async_trait>(
client: ConnectedClient<Self>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
Self: 'async_trait,

A client disconnected from the server.

Loading content...